+++ examples/client/unbanPeer.ts
@@ -7,7 +7,7 @@ import meow from 'meow';
 const cli = meow(
     `
   Usage
-    $ yarn ts-node <path-to-this-file> [options]
+    $ yarn run-example <path-to-this-file> [options]
 
   Required
     --ip, -i  The ip of the peer to connect to.