+++ packages/common/CHANGELOG.md
@@ -1,12 +1,16 @@
 # Changelog
 
-### Fixed
-
-- Cost calculation for `deployModule` transaction.
+## Unreleased
 
 ### Added
 
 - A `parseWallet` function to parse wallet export files
+
+## 7.0.1 2023-05-25
+
+### Fixed
+
+- Cost calculation for `deployModule` transaction.
 
 ## 7.0.0 2023-05-15